extract stats dtos
