+++ stats_models/src/main/java/ru/practicum/ewm/dto/HitDto.java
@@ -1,20 +1,20 @@
-package ru.practicum.ewm.stats;
+package ru.practicum.ewm.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
 import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.PositiveOrZero;
 
 @Data
 @Builder(toBuilder = true)
 @AllArgsConstructor
-public class StatsDto {
+public class HitDto {
     @NotBlank
     private String app;
     @NotBlank
     private String uri;
-    @PositiveOrZero
-    private Long hits;
+    @NotBlank
+    private String ip;
+    private String timestamp;
 }
