+++ stats_server/src/main/java/ru/practicum/ewm/hit/HitController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import ru.practicum.ewm.dto.HitDto;
 
 @Slf4j
 @RestController
