+++ stats_server/src/main/java/ru/practicum/ewm/hit/HitMapper.java
@@ -1,5 +1,7 @@
 package ru.practicum.ewm.hit;
 
+import ru.practicum.ewm.dto.HitDto;
+
 import javax.validation.ValidationException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
