+++ stats_server/src/main/java/ru/practicum/ewm/stats/StatsController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import ru.practicum.ewm.dto.StatsDto;
 
 import javax.validation.ValidationException;
 import java.time.LocalDateTime;
