+++ stats_server/src/test/java/ru/practicum/ewm/hit/HitControllerTest.java
@@ -11,6 +11,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
+import ru.practicum.ewm.dto.HitDto;
 
 import java.nio.charset.StandardCharsets;
